% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_db_tbl}
\alias{create_db_tbl}
\alias{create_db_tbl.tbl_sql}
\title{Create a table inside a database connection}
\usage{
create_db_tbl(x, name, type = c("view", "materialized view", "table"),
  .checks = TRUE, schema = "public")

\method{create_db_tbl}{tbl_sql}(x, name, type = c("view", "materialized view",
  "table"), .checks = TRUE, schema = "public")
}
\arguments{
\item{x}{An object representing the table to be created in the database}

\item{name}{The name of the table/view}

\item{type}{What type of view should be created in the database?}

\item{.checks}{Should we check for the existence of an object in the database
with `name` {name}?}

\item{schema}{schema}
}
\value{
Logical indicating whether the view was successfully created.
}
\description{
Create a table inside a database connection
}
