% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_db_tbls}
\alias{create_db_tbls}
\title{Create a table inside a database connection}
\usage{
create_db_tbls(..., type = c("view", "materialized view", "table"),
  .checks = TRUE, schema = "public")
}
\arguments{
\item{...}{Name-value pairs of expressions. The name will be the name of the
materialized view, and the expression should be the name of the `tbl_dbi`}

\item{type}{The type of table/view to create}

\item{.checks}{Perform checks to make sure materialized view doesn't already
exist?}

\item{schema}{database schema in which to place}
}
\description{
Create a table inside a database connection
}
