% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_schema}
\alias{create_schema}
\alias{create_schema.character}
\title{Create a new schema}
\usage{
create_schema(schema, db = NULL)

\method{create_schema}{character}(schema, db = NULL)
}
\arguments{
\item{schema}{name of schema}

\item{db}{(optional) A Database}
}
\description{
Create a new schema
}
