% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_views}
\alias{create_views}
\title{Create a materialized view from a lazy `tbl_dbi`}
\usage{
create_views(..., .checks = TRUE, schema = "public")
}
\arguments{
\item{...}{Name-value pairs of expressions. The name will be the name of the
materialized view, and the expression should be the name of the `tbl_dbi`}

\item{.checks}{Perform checks to make sure materialized view doesn't already
exist?}

\item{schema}{database schema in which to place}
}
\description{
Create a materialized view from a lazy `tbl_dbi`
}
