% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{dbCheckTable}
\alias{dbCheckTable}
\alias{dbCheckTable.src_dbi}
\alias{dbCheckTable.DBIConnection}
\title{Check to see if a Table/View exists in a DB source.}
\usage{
dbCheckTable(src, name, schema = "public")

\method{dbCheckTable}{src_dbi}(src, name, schema = "public")

\method{dbCheckTable}{DBIConnection}(src, name, schema = "public")
}
\arguments{
\item{src}{A data source / DB connection}

\item{name}{Name of table or view}

\item{schema}{Name of database schema in which to look}
}
\value{
TRUE if table does not exist; FALSE otherwise
}
\description{
If the table exists, prompts user for whether they want it dropped,
  with the additional option of cascading to any dependent views.
}
