% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{listMatViews}
\alias{listMatViews}
\alias{listMatViews.src_dbi}
\alias{listMatViews.DBIConnection}
\title{List all materialized views found in Postgres connection}
\usage{
listMatViews(src, schema = "public")

\method{listMatViews}{src_dbi}(src, schema = "public")

\method{listMatViews}{DBIConnection}(src, schema = "public")
}
\arguments{
\item{src}{A database connection (e.b. DBIConnection)}

\item{schema}{Name of database schema in which to look}
}
\value{
A character vector containing names of materialized views found.
}
\description{
List all materialized views found in Postgres connection
}
