% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{listTables}
\alias{listTables}
\alias{listTables.DBIConnection}
\title{List all tables found in the corresponding schema in the connection `src`}
\usage{
listTables(src, schema = "public")

\method{listTables}{DBIConnection}(src, schema = "public")
}
\arguments{
\item{src}{A data source}

\item{schema}{Name of database schema in which to look}
}
\value{
A character vector containing names of tablesfound.
}
\description{
List all tables found in the corresponding schema in the connection `src`
}
