% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{listViews}
\alias{listViews}
\alias{listViews.src_dbi}
\alias{listViews.DBIConnection}
\alias{listTables.src_dbi}
\title{List all (temporary) views found in Postgres connection}
\usage{
listViews(src, schema = "public")

\method{listViews}{src_dbi}(src, schema = "public")

\method{listViews}{DBIConnection}(src, schema = "public")

\method{listTables}{src_dbi}(src, schema = "public")
}
\arguments{
\item{src}{A data source}

\item{schema}{Name of database schema in which to look}
}
\value{
A character vector containing names of views found.
}
\description{
List all (temporary) views found in Postgres connection
}
