% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_base_tables}
\alias{load_base_tables}
\title{Load all tables into Global environment}
\usage{
load_base_tables(db = NULL)
}
\arguments{
\item{db}{(optional) A Database}
}
\description{
Load all tables into Global environment
}
