% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_db}
\alias{load_db}
\title{Load Database connections into Global Environment}
\usage{
load_db(dbname, config_file = NULL)
}
\arguments{
\item{dbname}{name of ODBC database source}

\item{config_file}{path of configuration file}
}
\value{
Copy of loaded database
}
\description{
Load Database connections into Global Environment
}
