% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_db_deprecated}
\alias{load_db_deprecated}
\title{Load Database connections into Global Environment}
\usage{
load_db_deprecated(name = NULL)
}
\arguments{
\item{name}{(OPTIONAL) name of ODBC database source}
}
\value{
Copy of loaded database
}
\description{
Load Database connections into Global Environment
}
