% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_materialized_views}
\alias{load_materialized_views}
\title{Load all materialized views into Global environment}
\usage{
load_materialized_views(db = NULL)
}
\arguments{
\item{db}{(optional) A Database}
}
\description{
Load all materialized views into Global environment
}
