% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_packages}
\alias{load_packages}
\title{Install or Load a list of Packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{a character vector of package names}
}
\description{
Install or Load a list of Packages
}
