% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_tables}
\alias{load_tables}
\title{Load all tables (all types) into Global environment}
\usage{
load_tables(db = NULL)
}
\arguments{
\item{db}{(optional) A Database}
}
\description{
Load all tables (all types) into Global environment
}
