% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{load_views}
\alias{load_views}
\title{Load all views into Global environment}
\usage{
load_views(db = NULL)
}
\arguments{
\item{db}{(optional) A Database}
}
\description{
Load all views into Global environment
}
