% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{refresh_materlialized_views}
\alias{refresh_materlialized_views}
\title{Refresh all materialized views}
\usage{
refresh_materlialized_views(db = NULL)
}
\arguments{
\item{db}{(optional) A Database}
}
\description{
Refresh all materialized views
}
