% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{script_path}
\alias{script_path}
\title{Get full path of the current script file}
\usage{
script_path()
}
\value{
The full path of current script file
}
\description{
Get full path of the current script file
}
