% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{sql_command}
\alias{sql_command}
\alias{sql_command.tbl_sql}
\title{Extract SQL query as string from lazy query.}
\usage{
sql_command(lazy_qry)

\method{sql_command}{tbl_sql}(lazy_qry)
}
\arguments{
\item{lazy_qry}{A `tbl_sql` object}
}
\value{
Character string representing SQL command.
}
\description{
Extract SQL query as string from lazy query.
}
