% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{tableTypes}
\alias{tableTypes}
\alias{tableTypes.src_dbi}
\alias{tableTypes.DBIConnection}
\title{Identify all Tables & Views In a Database Connection}
\usage{
tableTypes(src, schema = "public")

\method{tableTypes}{src_dbi}(src, schema = "public")

\method{tableTypes}{DBIConnection}(src, schema = "public")
}
\arguments{
\item{src}{A data source}

\item{schema}{Name of database schema in which to look}
}
\value{
A named character vector
}
\description{
Identify all Tables & Views In a Database Connection
}
